florida_church_agent/requirements.txt
beautifulsoup4==4.12.3
flask==3.0.3
openpyxl==3.1.5
pandas==2.2.2
playwright==1.54.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
rapidfuzz==3.9.6
requests==2.32.3
urllib3==2.2.2
pytest==8.3.2
pyinstaller==6.10.0
